pyproject.toml
[tool.poetry]
name = "django-migration-conflict-check"
version = "0.1.0"
description = ""
authors = ["daniel.lacko <daniel.lacko@remastr.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
atlassian-python-api = "^3.14.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
